+++ scripts/run_process_papers.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+# Script to run the paper processing script
+
+# Activate the virtual environment
+source /home/echo_/Code/asciimath/energy/.venv/bin/activate
+
+# Run the script
+python /home/echo_/Code/asciimath/energy/scripts/process_papers.py